Conflicts:
	Assets/Levels/RootIslandV1.world
	Assets/Models/LeafTree50Shape.dae
	Assets/Models/PaintedIslandBigShape.dae
	Assets/Models/PaintedIslandShape.dae
	Assets/Models/PaintedIslandShape1.dae
	Assets/Models/PaintedIslandShape2.dae
	Assets/Models/PaintedIslandShape3.dae
	Assets/Models/RootBridgeShape.dae
	Assets/Models/RootPlatformShape.dae
	Assets/Models/RootPlatformShape1.dae
	Assets/Models/RootPlatformShape3.dae
	Assets/Models/RootPlatformShape33.dae
	Assets/Models/RootPlatformShape35.dae
	Assets/Models/RootPlatformShape36.dae
	Assets/Models/RootPlatformShape7.dae
	Assets/Models/RootPlatformShape8.dae
	Assets/Models/RootsShape.dae
	Assets/Models/RootsShape1.dae
	Assets/Models/RootsShape2.dae
	Assets/Models/RootsShape3.dae
	Assets/Models/RootsShape4.dae
	Assets/Models/RuneGlowStone2Shape.dae
	Assets/Models/RuneGlowStone7Shape.dae
	Assets/Models/RuneGlowStoneShape.dae
	Assets/Models/waterShape.dae
	Assets/Shaders/pointlight.frag
	RootEngine/Render/Source/Texture.cpp

+++ RootSystems/Source/DataStructures/Quad.cpp
@@ -238,14 +238,14 @@ namespace RootForce
 		m_maxY = maxY;
 		m_minY = minY;
 
-		quadTreeBounds.m_maxX = RoundToPow2(maxX);
-		quadTreeBounds.m_minX = -RoundToPow2(abs(minX));
+		quadTreeBounds.m_maxX = (float)RoundToPow2(maxX);
+		quadTreeBounds.m_minX = (float)-RoundToPow2(abs(minX));
 
-		quadTreeBounds.m_maxY = maxY;
-		quadTreeBounds.m_minY = minY;
+		quadTreeBounds.m_maxY = (float)maxY;
+		quadTreeBounds.m_minY = (float)minY;
 
-		quadTreeBounds.m_maxZ = RoundToPow2(maxZ);
-		quadTreeBounds.m_minZ = -RoundToPow2(abs(minZ));
+		quadTreeBounds.m_maxZ = (float)RoundToPow2(maxZ);
+		quadTreeBounds.m_minZ = (float)-RoundToPow2(abs(minZ));
 
 		m_root = new QuadNode(quadTreeBounds);
 
@@ -339,15 +339,15 @@ namespace RootForce
 			int halfwidth = aabb->GetLengthX() / 2;
 			int halfheight = aabb->GetLengthZ() / 2;
 
-			Rectangle bottomLeft = Rectangle(aabb->m_minX, aabb->m_minZ, halfwidth, halfheight);
-			Rectangle bottomRight = Rectangle(aabb->m_minX + halfwidth, aabb->m_minZ, halfwidth, halfheight);
-			Rectangle topLeft = Rectangle(aabb->m_minX, aabb->m_minZ + halfheight, halfwidth, halfheight);
-			Rectangle topRight = Rectangle(aabb->m_minX + halfwidth, aabb->m_minZ + halfheight, halfwidth, halfheight);
+			Rectangle bottomLeft = Rectangle((int)aabb->m_minX, (int)aabb->m_minZ, halfwidth, halfheight);
+			Rectangle bottomRight = Rectangle((int)aabb->m_minX + halfwidth, (int)aabb->m_minZ, halfwidth, halfheight);
+			Rectangle topLeft = Rectangle((int)aabb->m_minX, (int)aabb->m_minZ + halfheight, halfwidth, halfheight);
+			Rectangle topRight = Rectangle((int)aabb->m_minX + halfwidth, (int)aabb->m_minZ + halfheight, halfwidth, halfheight);
 
-			AABB bottomLeftAABB = AABB(bottomLeft.m_x, bottomLeft.m_x + halfwidth, m_minY, m_maxY, bottomLeft.m_y, bottomLeft.m_y + halfheight);
-			AABB bottomRightAABB = AABB(bottomRight.m_x, bottomRight.m_x + halfwidth, m_minY, m_maxY, bottomRight.m_y, bottomRight.m_y + halfheight);
-			AABB topLeftAABB = AABB(topLeft.m_x, topLeft.m_x + halfwidth, m_minY, m_maxY, topLeft.m_y, topLeft.m_y + halfheight);
-			AABB topRightAABB = AABB(topRight.m_x, topRight.m_x + halfwidth, m_minY, m_maxY, topRight.m_y, topRight.m_y + halfheight);
+			AABB bottomLeftAABB = AABB((float)bottomLeft.m_x, (float)bottomLeft.m_x + (float)halfwidth, (float)m_minY, (float)m_maxY, (float)bottomLeft.m_y, (float)bottomLeft.m_y + (float)halfheight);
+			AABB bottomRightAABB = AABB((float)bottomRight.m_x, (float)bottomRight.m_x + (float)halfwidth, (float)m_minY, (float)m_maxY, (float)bottomRight.m_y, (float)bottomRight.m_y + (float)halfheight);
+			AABB topLeftAABB = AABB((float)topLeft.m_x, (float)topLeft.m_x + (float)halfwidth, (float)m_minY, (float)m_maxY, (float)topLeft.m_y, (float)topLeft.m_y + (float)halfheight);
+			AABB topRightAABB = AABB((float)topRight.m_x, (float)topRight.m_x + (float)halfwidth, (float)m_minY, (float)m_maxY, (float)topRight.m_y, (float)topRight.m_y + (float)halfheight);
 
 			std::vector<Polygon> tr;
 			std::vector<Polygon> tl;
